Allow progress to be set in between frames (#1261)
Fixes #1250

+++ lottie/src/main/java/com/airbnb/lottie/utils/LottieValueAnimator.java
@@ -146,7 +146,7 @@ public class LottieValueAnimator extends BaseLottieAnimator implements Choreogra
     setFrame((int) frame);
   }
 
-  public void setFrame(int frame) {
+  public void setFrame(float frame) {
     if (this.frame == frame) {
       return;
     }
